Код разделен на модули, создана точка входа
